+++ CHANGELOG.md
@@ -7,6 +7,7 @@ description: |-
 
 # VKCS Provider's changelog
 #### v0.4.2 (unreleased)
+- Add data source for images
 - Add options to unzip and decompress downloaded image in Images image resource
 - Fix leaving a damaged file in cache when an error ocurred during downloading an image from source
 - Fix schedule.time format of vkcs_backup_plan datasource